requirements.txt
Django==1.4.3
Markdown==2.2.1
argparse==1.2.1
djangorestframework==2.1.17
wsgiref==0.1.2
